# BACKEND/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CredPal technical assessment project (Lumen/vue)",
        "contact": {
            "name": "Idisimagha Dublin-green (Cliquedom Technologies)",
            "email": "greendublin007@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/apiv1/getAccountByAccountNumber/{accountNumber}": {
            "get": {
                "tags": [
                    "getAccountByAccountNumber"
                ],
                "summary": "Get account details by account number.",
                "description": "Get account details by account number",
                "operationId": "/apiv1/getAccountByAccountNumber/{accountNumber}",
                "parameters": [
                    {
                        "name": "accountNumber",
                        "in": "path",
                        "description": "accountNumber to get account details",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns some account data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Token not provided.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/apiv1/users/": {
            "get": {
                "tags": [
                    "GetAllUsers"
                ],
                "operationId": "/apiv1/users/",
                "responses": {
                    "200": {
                        "description": "Returns some users data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Token not provided.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/createUser/": {
            "post": {
                "tags": [
                    "CreateUser"
                ],
                "summary": "Create a user and create a new wallet account.",
                "description": "Create a user and create a new wallet account. Depending on configuration sends new registration SMS.",
                "operationId": "/auth/createUser/",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "schema": "PostRequest",
                                "title": "PostRequest",
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password",
                                    "referrer_id",
                                    "mobile"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "mobile": {
                                        "type": "integer"
                                    },
                                    "referrer_id": {
                                        "type": "string",
                                        "default": "0"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns some users data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "user created!"
                                        },
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "first_name": "first-name",
                                                "last_name": "last-name",
                                                "email": "email@gmail.com",
                                                "mobile": "08095060650"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Returns some users data but does not create new user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "unable to create user, something went wrong."
                                        },
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "first_name": "first-name",
                                                "last_name": "last-name",
                                                "email": "email@gmail.com",
                                                "mobile": "08095060650"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer Token for authentication",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}